+++ src/threading/core.clj
@@ -7,24 +7,33 @@
             [shuriken.debug :refer [debug-print]]
             [shuriken.associative :refer [map-keys map-vals]]))
 
-(s/def ::macro-variants
+(s/def ::threading-variant
+  (s/cat
+    :arrow     symbol?
+    :docstring (s/? string?)
+    :opts      (s/? map?)))
+
+(s/def ::threading-variants
   (s/and vector?
-         (s/* (s/cat
-                :macro     symbol?
-                :docstring (s/? string?)
-                :opts      (s/? map?)))))
+         (s/* ::threading-variant)))
 
 (s/def ::defthreading-args
   (s/cat
-    :name-prefix (s/? symbol?)
-    :doc-prefix  (s/? string?)
-    :variants    ::macro-variants
-    :bodies      :shuriken.spec/args+bodies))
+    :thread-name    (s/? symbol?)
+    :arrow-position (s/? #{:prefix :suffix})
+    :doc-prefix     (s/? string?)
+    :variants       ::threading-variants
+    :bodies         :shuriken.spec/args+bodies))
 
-;; TODO: document
-(defmacro ^{:arglists [[name-prefix? doc-prefix? [(variant docstring? opts?)+]
-                        [params*] body]
-                       [name-prefix? doc-prefix? [(variant docstring? opts?)+]
+(defmacro ^{:arglists [[thread-name? arrow-position?
+                        doc-prefix?
+                        [(variant docstring? opts?)+]
+                        [params*]
+                        body*]
+                       [thread-name? arrow-position?
+                        doc-prefix?
+                        [(variant docstring? opts?)+]
+                        [params*]
                         ([params*] prepost-map? body)+]]}
   defthreading
   "Defines a threading arrow with variants.
@@ -32,7 +41,7 @@
   Each variant is defined by a sequence like `(name docstring? opts?)`.
 
   The name of each defined threading arrow will consist in the
-  variant symbol (usually `->` and `->>`) prefixed by `name-prefix`
+  variant symbol (usually `->` and `->>`) prefixed by `thread-name`
   if provided.
   Same goes for each arrow docstring: it will be prefixed with
   `doc-prefix` if provided.
@@ -73,28 +82,37 @@
   ;; 1 : perform some side-effect.
   ;; => 1
   ```
-  "
+
+  Remarks:
+
+  - `name` is optional.
+  - `name` must be followed by the `:prefix` `arrow-position` to
+    define fletchings."
   [& args]
-  (let [{:keys [name-prefix doc-prefix variants bodies]}
+  (let [{:keys [thread-name arrow-position doc-prefix variants bodies]}
         (conform! ::defthreading-args args)]
-    (cons 'do (for [{:keys [macro docstring opts]} variants
-                    :let [specific-name (symbol (str name-prefix macro))
+    (cons 'do (for [{:keys [arrow docstring opts]} variants
+                    :let [specific-name (-> (case (or arrow-position :suffix)
+                                              :suffix (str thread-name arrow)
+                                              :prefix (str arrow thread-name))
+                                            symbol)
                           specific-doc
                           (if (or doc-prefix docstring)
                             (clojure.string/trim
-                              (str doc-prefix \newline
-                                   (->> docstring
-                                        lines
-                                        (map #(-> (str/replace % #"^\s+" "")
-                                                  (->> (str "  "))))
-                                        join-lines))))
-                          bodies (map (fn [{:keys [body] :as plan}]
-                                        (->> body)
-                                        (assoc plan :body
-                                          `((let [~'&threading-variant '~macro
-                                                  ~'&threading-opts '~opts]
-                                              ~@body))))
-                                      bodies)]]
+                              (str doc-prefix
+                                   (some->> docstring
+                                            lines
+                                            (map #(-> (str/replace % #"^\s+" "")
+                                                      (->> (str "  "))))
+                                            join-lines
+                                            (str \newline)))))
+                          bodies
+                          (map (fn [{:keys [body] :as plan}]
+                                 (assoc plan :body
+                                   `((let [~'&threading-variant '~arrow
+                                           ~'&threading-opts '~opts]
+                                       ~@body))))
+                               bodies)]]
                 (do (->> {:op        'defmacro
                           :name       specific-name
                           :doc-string specific-doc
@@ -105,49 +123,54 @@
                          (s/unform :shuriken.spec/defmacro-form)))))))
 
 (defthreading
-  "A threading arrow fletching used to change the position at which the
-  threaded form will be injected in the threading slots of the threading form."
-  [>-  "Transitions from a thread-last to a thread-first threading style."
-   >>- "Transitions from a thread-first to a thread-last threading style."]
-  [first-expr last-expr]
+  "A threading fletching used to change the position at which the threaded
+  form will be injected in the threading slots of the threading form."
+  [>-  "Transitions from a thread-first to a thread-last threading style."
+   >>- "Transitions from a thread-last to a thread-first threading style."]
+  [first-expr second-expr]
   (case &threading-variant
-    >-  (let [threaded-expr last-expr
+    >-  (let [threaded-expr first-expr
+              [verb & args] second-expr]
+          `(~verb ~@args ~threaded-expr))
+    >>- (let [threaded-expr second-expr
               [verb & args] first-expr]
-          `(~verb ~threaded-expr ~@args))
-    >>- (let [threaded-expr first-expr
-              [verb & args] last-expr]
-          `(~verb ~@args ~threaded-expr))))
+          `(~verb ~threaded-expr ~@args))))
 
-(defmacro >-> [expr threaded-expr]
-  `(>- (-> ~expr) ~threaded-expr))
+(defmacro >->
+  "Shorthand for `(>- (-> ...)."
+  [threaded-expr expr]
+  `(>- ~threaded-expr (-> ~expr)))
 
-(defmacro >->> [expr threaded-expr]
-  `(>- (->> ~expr) ~threaded-expr))
+(defmacro >->>
+  "Shorthand for `(>- (->> ...)."
+  [threaded-expr expr]
+  `(>- ~threaded-expr (->> ~expr)))
 
-(defmacro >>-> [threaded-expr expr]
-  `(>>- ~threaded-expr (-> ~expr)))
+(defmacro >>->
+  "Shorthand for `(>>- (-> ...)."
+  [expr threaded-expr]
+  `(>>- (-> ~expr) ~threaded-expr))
 
-(defmacro >>->> [threaded-expr expr]
-  `(>>- ~threaded-expr (->> ~expr)))
+(defmacro >>->>
+  "Shorthand for `(>>- (->> ...)."
+  [expr threaded-expr]
+  `(>>- (->> ~expr) ~threaded-expr))
 
 (defthreading
-  [<-
-   "Used to provide arbitrary input and output values to forms threading in the
-   style of `->`.
-   The threaded form will be evaluated although its value will be
-   discarded."
-
-   <<-
-   "Like [[<-]] but must be used in the context of a form threading in the
-   style of `->>`."]
-  [& body]
+  "As an equivalent to `constantly`, the anti-threading arrow is used to
+  return an arbitrary value from a threading slot.
+  The threaded expression will be evaluated although its value will be
+  discarded."
+  [<-  "To be used in the context of a form threading in the style of `->`."
+   <<- "To be used in the context of a form threading in the style of `->>`."]
+  [first-expr second-expr]
   (case &threading-variant
-    <-  `(do ~@body)
-    <<- `(do ~(last body) ~@(butlast body))))
+    <-  `(do ~second-expr)
+    <<- `(do ~first-expr)))
 
 (defthreading tap
-  "Threads the expr through the forms then returns the value of
-  the initial expr."
+  "Threads `expr` through `forms` then returns the value of
+  the initial expression."
   [->  "Threads like `->`."
    ->> "Threads like `->>`."]
   [expr & forms]
@@ -157,19 +180,22 @@
 
 (defmacro tap
   "Evaluates expressions in order returning the value of the first.
-  Will thread the first expr into any subsequent threading expr.
+  Will thread the first expression into any immediate threading
+  subexpressions.
 
+  ```clojure
   (tap 123
        (println \"yo\")
        (some-> inc println))
   ; yo
   ; 124
-  ; => 123"
+  ; => 123
+  ```"
   [expr & body]
   (let [result-sym (gensym "result-")]
     `(let [~result-sym ~expr]
        ~@(map (fn [x]
-                (if (some->> x first str (re-matches #"^.*->>?$"))
+                (if (some->> x first str (re-matches #"^.*->>?|-••?$"))
                   `(-> ~result-sym ~x)
                   x))
               body)
@@ -361,7 +387,7 @@
                expr#)))
 
 (defthreading juxt
-  "Threads `expr` through each form individually collecting the results in a
+  "Threads `expr` through each form individually, collecting the results in a
   lazy sequence."
   [->  "Threads like `->`."
    ->> "Threads like `->>`."]
@@ -375,7 +401,7 @@
 
 (defthreading let
   "Threads `expr` through each bound expression then through each form in
-  `body`. Returns the value of its last expr."
+  `body` like `let`."
   [->  "Threads like `->`."
    ->> "Threads like `->>`."]
   [expr bindings & body]
@@ -384,7 +410,69 @@
     `(as-> ~expr ~input-sym
        (let ~(->> bindings
                   (partition 2)
-                  (>- (map-vals-> adapt))
+                  (>>- (map-vals-> adapt))
                   (apply concat)
                   vec)
          ~@(map adapt body)))))
+
+(defthreading binding
+  "Threads `expr` through each bound expression then through each form in
+  `body` like `binding`."
+  [->  "Threads like `->`."
+   ->> "Threads like `->>`."]
+  [expr bindings & body]
+  (let [input-sym (gensym "input-")
+        adapt (fn [expr] `(~&threading-variant ~input-sym ~expr))]
+    `(as-> ~expr ~input-sym
+       (binding ~(->> bindings
+                      (partition 2)
+                      (>>- (map-vals-> adapt))
+                      (apply concat)
+                      vec)
+         ~@(map adapt body)))))
+
+(def ^:private •-value-sym (gensym "•-value-"))
+
+(defthreading
+  "Stores `expr` on the stack and returns the computation represented by
+  `form`. In `form`, any deep occurence of the `-•` arrow will thread
+  this value stored on the stack to its inner expressions.
+
+  Note that `-•` can only be used within the body of a `•-` form."
+  [•-  "Must be used in the context of a thread-first arrow."
+   ••- "Must be used in the context of a thread-last arrow."]
+  [first-expr second-expr]
+  (let [[expr form] (case &threading-variant
+                      •-  [first-expr  second-expr]
+                      ••- [second-expr first-expr])]
+    `(let [~•-value-sym ~expr]
+       ~form)))
+
+(defthreading
+  "See [[•-]]."
+  [-•  "Threads like `->`."
+   -•• "Threads like `->>`."]
+  [& forms]
+  (let [arrow (case &threading-variant
+                -• '->
+                -•• '->>)]
+    `(~arrow ~•-value-sym ~@forms)))
+
+(defthreading args :prefix
+  "Fletching that uses the arrow in the threading form to thread the
+  expression in all subforms of its subform. `form` must have only one subform.
+
+  Fletching that moves the arrow in the threading form inside its subform's
+  arguments. The threading arrow must have only one subform."
+  [>-  "Will thread the expr into the subforms with `->`."
+   >>- "Will thread the expr into the subforms with `->>`."]
+  [first-expr second-expr]
+  (let [[expr [arrow subform]] (case &threading-variant
+                                 >-  [first-expr  second-expr]
+                                 >>- [second-expr first-expr])
+        [head & subform-subforms] subform
+        v-sym (gensym "v-")]
+    `(let [~v-sym ~expr]
+       (~head ~@(map (fn [sub]
+                       `(~arrow ~v-sym ~sub))
+                     subform-subforms)))))